+++ sphinx/source/changelog.rst
@@ -15,6 +15,9 @@ Changelog
 * Added the ``polarity`` simulated LOAD configuration parameter to
   :ref:`tap2sna.py <tap2sna-sim-load>` (to specify the EAR bit reading produced
   by the first pulse on the tape)
+* Added the ``read-in-r-c`` simulated LOAD configuration parameter to
+  :ref:`tap2sna.py <tap2sna-sim-load>` (to specify whether to yield a simulated
+  port reading for 'IN r,(C)' instructions)
 * The output snapshot argument of :ref:`tap2sna.py` is now optional
 * Added support to :ref:`tap2sna.py <tap2sna-conf>` for register pairs
   (``r[bc]``, ``r[de]`` etc.) in the ``TraceLine`` configuration parameter